# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "App"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/profile": {
      "get": {
        "operationId": "AuthController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/patient/login": {
      "post": {
        "operationId": "MobileAuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MobileLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns JWT Token"
          },
          "401": {
            "description": "Invalid PIN or ID"
          }
        },
        "summary": "Login for Mobile App using PIN",
        "tags": [
          "Mobile Auth"
        ]
      }
    },
    "/patients": {
      "post": {
        "operationId": "PatientsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePatientDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Patients"
        ]
      },
      "get": {
        "operationId": "PatientsController_findAll",
        "parameters": [
          {
            "name": "search",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Patients"
        ]
      }
    },
    "/patients/{id}": {
      "get": {
        "operationId": "PatientsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Patients"
        ]
      },
      "patch": {
        "operationId": "PatientsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePatientDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Patients"
        ]
      },
      "delete": {
        "operationId": "PatientsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Patients"
        ]
      }
    },
    "/patients/{id}/routines": {
      "get": {
        "operationId": "PatientsController_getPatientRoutines",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Patients"
        ]
      }
    },
    "/patients/{id}/access-code": {
      "post": {
        "operationId": "PatientsController_generateAccessCode",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Patients"
        ]
      },
      "delete": {
        "operationId": "PatientsController_revokeAccessCode",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Patients"
        ]
      }
    },
    "/patients/{id}/access-code/status": {
      "get": {
        "operationId": "PatientsController_hasAccessCode",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Patients"
        ]
      }
    },
    "/routines": {
      "post": {
        "operationId": "RoutinesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoutineDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Routines"
        ]
      },
      "get": {
        "operationId": "RoutinesController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Routines"
        ]
      }
    },
    "/routines/{id}": {
      "get": {
        "operationId": "RoutinesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Routines"
        ]
      },
      "put": {
        "operationId": "RoutinesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoutineDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Routines"
        ]
      },
      "delete": {
        "operationId": "RoutinesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Routines"
        ]
      }
    },
    "/routines/{id}/clone": {
      "post": {
        "operationId": "RoutinesController_clone",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Routines"
        ]
      }
    },
    "/exercises": {
      "get": {
        "operationId": "ExercisesController_findAll",
        "parameters": [
          {
            "name": "search",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Exercises"
        ]
      },
      "post": {
        "operationId": "ExercisesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExerciseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Exercises"
        ]
      }
    },
    "/exercises/{id}": {
      "get": {
        "operationId": "ExercisesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Exercises"
        ]
      },
      "patch": {
        "operationId": "ExercisesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateExerciseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Exercises"
        ]
      },
      "delete": {
        "operationId": "ExercisesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Exercises"
        ]
      }
    },
    "/routines/{id}/stats": {
      "get": {
        "operationId": "AnalyticsController_getRoutineStats",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Analytics"
        ]
      }
    },
    "/routines/{id}/history": {
      "get": {
        "operationId": "AnalyticsController_getRoutineHistory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Analytics"
        ]
      }
    },
    "/dashboard/stats": {
      "get": {
        "operationId": "DashboardController_getStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Dashboard"
        ]
      }
    },
    "/mobile/routine/active": {
      "get": {
        "operationId": "MobileController_getActiveRoutine",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns full routine object with exercises",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActiveRoutineResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get active routine for logged-in patient",
        "tags": [
          "Mobile Features"
        ]
      }
    },
    "/sessions": {
      "post": {
        "operationId": "SessionsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new therapy session record",
        "tags": [
          "Sessions"
        ]
      }
    }
  },
  "info": {
    "title": "Symma Platform API",
    "description": "API documentation for Web Dashboard and Mobile App",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {}
      },
      "MobileLoginDto": {
        "type": "object",
        "properties": {}
      },
      "CreatePatientDto": {
        "type": "object",
        "properties": {}
      },
      "UpdatePatientDto": {
        "type": "object",
        "properties": {}
      },
      "CreateRoutineDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateRoutineDto": {
        "type": "object",
        "properties": {}
      },
      "CreateExerciseDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateExerciseDto": {
        "type": "object",
        "properties": {}
      },
      "ActiveRoutineExerciseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "keyName": {
            "type": "string"
          },
          "description": {
            "type": "object"
          },
          "type": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "mobileModule": {
            "type": "object",
            "description": "Mobile module for strategy selection (SMILE, BROWS, JAW, KISS, EYES, EYES_INVERSE)"
          },
          "assetAnimationUrl": {
            "type": "object"
          },
          "assetTutorialVideoUrl": {
            "type": "object"
          }
        },
        "required": [
          "id",
          "name",
          "keyName",
          "type",
          "category"
        ]
      },
      "ActiveRoutineItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "orderIndex": {
            "type": "number"
          },
          "sets": {
            "type": "number",
            "description": "Number of sets for this exercise"
          },
          "repsPerSet": {
            "type": "number",
            "description": "Number of repetitions per set"
          },
          "targetHoldSeconds": {
            "type": "number",
            "description": "Hold time in seconds (0 for isotonic exercises)"
          },
          "restBetweenSets": {
            "type": "number",
            "description": "Rest time between sets in seconds"
          },
          "difficultyLevel": {
            "type": "number",
            "description": "Difficulty level multiplier (1.0 = normal)"
          },
          "strictMode": {
            "type": "boolean",
            "description": "Whether strict mode is enabled"
          },
          "exercise": {
            "$ref": "#/components/schemas/ActiveRoutineExerciseDto"
          }
        },
        "required": [
          "id",
          "orderIndex",
          "sets",
          "repsPerSet",
          "targetHoldSeconds",
          "restBetweenSets",
          "difficultyLevel",
          "strictMode",
          "exercise"
        ]
      },
      "ActiveRoutineResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "startDate": {
            "type": "string"
          },
          "endDate": {
            "type": "object"
          },
          "status": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActiveRoutineItemDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "startDate",
          "status",
          "items"
        ]
      },
      "SessionItemDto": {
        "type": "object",
        "properties": {
          "exerciseId": {
            "type": "string"
          },
          "repsCompleted": {
            "type": "number"
          },
          "difficulty": {
            "type": "number",
            "default": 0
          },
          "averageAccuracy": {
            "type": "number",
            "description": "Average symmetry score (0-100)"
          },
          "seriesData": {
            "type": "object",
            "description": "Detailed series data"
          }
        },
        "required": [
          "exerciseId",
          "repsCompleted"
        ]
      },
      "CreateSessionDto": {
        "type": "object",
        "properties": {
          "routineId": {
            "type": "string"
          },
          "startTime": {
            "type": "string"
          },
          "endTime": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SessionItemDto"
            }
          }
        },
        "required": [
          "routineId",
          "startTime",
          "endTime",
          "items"
        ]
      }
    }
  }
}